brew audit is just annoying
brew audit just produces silly errors like "excess whitespace on line 6" and "you can't depend on doxygen --with-graphviz because I said so".

It's not needed. If all we're doing from now on is updating the version number and the hash then there shouldn't be anything wrong with the formula. This also means that the script can be run on linux and macos without issues.