Update test for v3
The fact that this test doesn't run every time the package is installed probably proves that it doesn't need to exist. Homebrew recommends that each formula have a test but this forumula is not in the core tap so we don't really need to abide by any rules.

I'm keeping it anyway because it's not doing any harm and is somewhat useful at times.